requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
click==8.0.3; python_version >= '3.6'
dominate==2.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
flask-bootstrap4==4.0.2
flask==2.0.2
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.3; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
prometheus-client==0.12.0
uwsgi==2.0.20
visitor==0.1.3
werkzeug==2.0.2; python_version >= '3.6'
